backend/requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
marshmallow-sqlalchemy==1.1.0
